Fix comment undefined error, Implemented Notification (#36)
 implemented Notification

+++ src/app/components/comment-list/comment-list.component.ts
@@ -2,38 +2,44 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import {CommentService} from '../../services/comment.service';
 import {TaskService} from '../../services/task.service';
-import {FormBuilder} from '@angular/forms';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {UserService} from '../../services/user.service';
 import {Comment} from '../../model/Comment';
 import {User} from "../../model/User";
+import {Task} from "../../model/Task";
+import {DialogService} from "../../services/dialog.service";
+import {NotificationService} from "../../services/notification.service";
 
 @Component({
   selector: 'app-comment-list',
   templateUrl: './comment-list.component.html',
   styleUrls: ['./comment-list.component.scss']
 })
+
 export class CommentListComponent implements OnInit {
-  comments: Comment[];
-  allComments: Comment[];
-  task;
-  checkoutForm;
+
+  comments: Comment[] = [];
+  allComments: Comment[] = [];
+  task:Task = null;
+  createCommentForm:FormGroup;
   user:User;
   isSuccesfullMessage: boolean;
   isValidationMessage:boolean;
   searchText:string;
   isMoreView:boolean;
 
-
   constructor(private commentService: CommentService, private userService: UserService, private formBuilder: FormBuilder,
               private taskService: TaskService,
-              private actRouter: ActivatedRoute, private router: Router) {
-
+              private actRouter: ActivatedRoute, private router: Router,private dialogService: DialogService, private notificationService : NotificationService) {
+    this.createCommentForm = this.formBuilder.group({
+      description: ['', [Validators.required]],
+    });
 
   }
 
   ngOnInit() {
     this.taskService.findTaskById(this.actRouter.snapshot.paramMap.get('id'))
-      .subscribe(data => {
+      .subscribe((data:Task) => {
         this.task = data;
       })
 
@@ -43,7 +49,7 @@ export class CommentListComponent implements OnInit {
           this.comments = data.slice(0,10);
       });
 
-    this.checkoutForm = this.formBuilder.group({
+    this.createCommentForm = this.formBuilder.group({
       description: ''
     });
     this.userService.getAuthenticatedUser().subscribe((date:User) => {
@@ -51,16 +57,25 @@ export class CommentListComponent implements OnInit {
     });
 
   }
-
+  get description() {
+    return this.createCommentForm.get('description');
+  }
   handleEdit(comment:Comment) {
     this.router.navigateByUrl('/home/comments/edit/' + comment.id);
   }
 
   handleDelete(comment:Comment) {
-    this.commentService.deleteComment(comment.id).subscribe(date =>{
-      this.comments = this.comments.filter(item => item.id !== comment.id);
-      this.allComments = this.allComments.filter(item => item.id !== comment.id);
-      console.log(date);
+    this.dialogService.openConfirmDialog('Are you sure to delete this comment?')
+      .afterClosed().subscribe(res => {
+      if (res) {
+        this.commentService.deleteComment(comment.id).subscribe(date => {
+          this.allComments = this.allComments.filter(item => item.id !== comment.id);
+          this.comments = this.allComments.slice(0,10);
+          if(this.comments.length > 10) this.isMoreView = false;
+          console.log(date);
+          this.notificationService.showSuccessWithTimeout("Comment has been successfully deleted!","Success",3200);
+        });
+      }
     });
 
   }
@@ -73,8 +88,9 @@ export class CommentListComponent implements OnInit {
       ((date: Comment) => {
           this.isSuccesfullMessage = true;
           this.isValidationMessage = false;
-          this.checkoutForm.controls['description'].setValue('');
+          this.createCommentForm.controls['description'].setValue('');
           console.log(comment);
+          this.notificationService.showSuccessWithTimeout("Comment has been successfully created!","Success",3200);
           this.commentService.findCommentById(date.id).subscribe((result: Comment) => {
             this.comments.unshift(result);
             this.comments = this.comments.slice(0,10);
@@ -85,24 +101,25 @@ export class CommentListComponent implements OnInit {
     }
     else {
       this.isValidationMessage = true;
+      this.notificationService.showErrorWithTimeout("You do not create the comment with the empty description!!!","Error",4200);
     }
   }
 
 
   viewMore(){
-
     this.comments = this.allComments.slice(0,this.comments.length+10);
-
-
   }
+
   messageSuccsesfulClose(){
     this.isSuccesfullMessage= false;
   }
-  messageValidationClose(){
-    this.isValidationMessage = false;
-  }
-  checkOnOwner(comment:Comment) {
-        return comment.userId == this.user.id;
+
+
+  checkOnOwner(comment:Comment):boolean {
+    if(this.user) {
+      return comment.userId === this.user.id;
+    }
+    return false;
   }
 
   goBackToTask(){
@@ -120,6 +137,7 @@ export class CommentListComponent implements OnInit {
       return -1;
     })
     this.comments = this.allComments.slice(0,10);
+    this.notificationService.showInfoWithTimeout("The comment list was sorted by date, first new then old","Success",4200);
   }
   sortOld(){
     this.allComments.sort((a, b) => {
@@ -132,14 +150,14 @@ export class CommentListComponent implements OnInit {
       return -1;
     })
     this.comments = this.allComments.slice(0,10);
+    this.notificationService.showInfoWithTimeout("The comment list was sorted by date, first old then new","Success",4200);
+
   }
 
   search(){
-
       this.comments = this.allComments.filter(value =>
         value.userId.toString().indexOf(this.searchText) > -1
       );
-
       if(this.searchText !== ''){
         this.isMoreView = true;
       }
@@ -150,6 +168,11 @@ export class CommentListComponent implements OnInit {
         }
       }
   }
+  messageValidationClose(){
+    this.isValidationMessage = false;
+  }
+
+
 }
